+++ src/main/java/blackjack/BlackjackApplication.java
@@ -1,7 +1,5 @@
 package blackjack;
 
-import blackjack.controller.BlackjackGame;
-
 public class BlackjackApplication {
 
     public static void main(String[] args) {
